+++ src/App.js
@@ -4,6 +4,7 @@ import Layout from './components/Layout/Layout';
 import Login from './components/Login/Login';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Game from './components/Game/game';
+import Scoreboard from './components/Scoreboard/Scoreboard';
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
         <Layout>
           <Switch>
             <Route path="/game" component={Game}></Route>
+            <Route path="/scoreboard" component={Scoreboard}></Route>
             <Route path="/" exact component={Login}></Route>
           </Switch>
         </Layout>
