+++ src/components/Game-Header/GameHeader.css
@@ -1,0 +1,27 @@
+.main-header {
+    color:red;
+    display: flex;
+    flex-direction: column;
+    font-size: 1.5rem;
+    width: 90%;
+}
+.user {
+    display: flex;
+    flex-grow: 1;
+    text-transform: uppercase;
+    font-family: 'Montserrat', sans-serif;
+    line-height: 1.5;
+}
+.user-icon {
+    padding-right: 1%;
+}
+.fast-fingers {
+    font-size: 2rem;
+}
+.header-name {
+    display: flex;
+    flex-direction: row;
+}
+.score {
+    text-transform: uppercase;
+}
