+++ src/components/Main-Game/ScoreBoard.css
@@ -1,0 +1,12 @@
+.title {
+    text-align: center;
+    font-size: 18px;
+}
+.game-score {
+    align-self: initial;
+    color: white;
+    padding-top: 3%;
+}
+.game-individual {
+    line-height: 2.1;
+}
